Fixed broken #find.
Mongoid::Slug::Criterion was not being included into Mongoid::Criteria correctly, and as a result no slug aware #find method was implemented. There were no specs testing this.

This commit removes that module/include pattern and replaces it with direct extension of the Mongoid::Criteria class. The #find method will now attempt to determine whether it is being supplied with _ids or slugs and responds accordingly.