Refactored Mongoid::Criteria#find.
Makes it much less likely to break if the Mongoid implementation of same changes.

Removed the optional hash which makes it possible to force finding by slug or id. Replaced with a #find_by_slug method. Reinstated method of same name in Document to enable the user to explicitly find by id (required in the event that the document uses Strind _ids.